+++ source/components/Overlay.css
@@ -5,6 +5,7 @@
 	right: 0;
 	bottom: 0;
 	background: rgba(255, 255, 255, 0.9);
+	overflow: auto;
 	z-index: 1;
 }
 #overlayContent {
