+++ source/components/Overlay.js
@@ -1,9 +1,17 @@
 import FocusTrap from 'focus-trap-react'
-import React from 'react'
+import React, { useEffect } from 'react'
 import * as animate from 'Ui/animate'
 import { LinkButton } from 'Ui/Button'
 import './Overlay.css'
 export default function Overlay({ onClose, children, ...otherProps }) {
+	useEffect(() => {
+		const body = document.getElementsByTagName('body')[0]
+		body.classList.add('no-scroll');
+		return () => {
+			body.classList.remove('no-scroll')
+		}
+	}
+		, [])
 	return (
 		<div id="overlayWrapper">
 			<animate.fromBottom>
