+++ source/sites/mon-entreprise.fr/layout/Footer/Footer.js
@@ -20,20 +20,20 @@ type OwnProps = {}
 const feedbackBlacklist = [
 	['index'],
 	['entreprise', 'statutJuridique', 'index'],
-	['sécuritéSociale', 'indépendant'],
-	['sécuritéSociale', 'auto-entrepreneur'],
-	['sécuritéSociale', 'assimilé-salarié'],
-	['sécuritéSociale', 'salarié']
+	['simulateurs', 'indépendant'],
+	['simulateurs', 'auto-entrepreneur'],
+	['simulateurs', 'assimilé-salarié'],
+	['simulateurs', 'salarié']
 ].map(lensPath)
 
 const Footer = ({ sitePaths }) => {
 	const hrefLink =
 		hrefLangLink[i18n.language][
-			decodeURIComponent(
-				(process.env.NODE_ENV === 'production'
-					? window.location.protocol + '//' + window.location.host
-					: '') + window.location.pathname
-			).replace(/\/$/, '')
+		decodeURIComponent(
+			(process.env.NODE_ENV === 'production'
+				? window.location.protocol + '//' + window.location.host
+				: '') + window.location.pathname
+		).replace(/\/$/, '')
 		] || []
 	return (
 		<div className="footer-container">
@@ -108,8 +108,8 @@ const Footer = ({ sitePaths }) => {
 								) : hrefLang === 'en' ? (
 									<> Switch to English {emoji('🇬🇧')}</>
 								) : (
-									hrefLang
-								)}
+											hrefLang
+										)}
 							</a>
 						))}
 					</p>
