--- source/sites/mon-entreprise.fr/layout/Navigation/Navigation.css
@@ -1,118 +1,0 @@
-.navigation .progress {
-	height: 2px;
-	position: absolute;
-	bottom: 5px;
-	background: rgba(41, 117, 209, 0.2);
-	border-radius: 0.1rem;
-	width: 100%;
-}
-
-.navigation .progress .bar {
-	transition: width 0.3s;
-	background-color: rgb(41, 117, 209);
-	height: 100%;
-}
-.navigation__container {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-end;
-	flex-shrink: 0;
-	min-height: 100%;
-	background-color: rgb(227, 237, 249);
-	background-color: var(--lighterColour);
-}
-
-.navigation {
-	padding: 1rem 2rem 3rem 0.6rem;
-	flex: 1;
-	min-width: 20rem;
-}
-
-.navigation ul {
-	list-style: none;
-	margin: 0;
-	padding-left: 1.2rem;
-}
-.navigation li {
-	padding: 0;
-}
-.navigation > ul > li > ul {
-	visibility: visible !important;
-	height: auto !important;
-}
-.navigation__item:not(.opened) + ul {
-	visibility: hidden;
-	height: 0;
-}
-.navigation__item:hover {
-	opacity: 0.8;
-}
-.navigation__item:hover * {
-	opacity: 1 !important;
-}
-
-.navigation__caret {
-	width: 30px;
-	text-align: center !important;
-	transition: transform 0.2s;
-	font-size: 120% !important;
-}
-.navigation__item.opened .navigation__caret {
-	transform: rotate(90deg) translateX(1px) translateY(-1px);
-}
-.navigation__item {
-	left: -30px;
-	white-space: nowrap;
-	position: relative;
-}
-.navigation__item.active {
-	color: rgb(41, 117, 209);
-	color: var(--colour);
-}
-
-.navigation li > :first-child {
-	position: relative;
-	display: inline-block;
-	text-decoration: none;
-	width: 100%;
-	padding: 0.2rem 0;
-}
-.navigation > ul > li ul > li > :first-child:not(.active):not(:hover) {
-	opacity: 0.6;
-}
-.navigation a:not(.active) {
-	color: inherit;
-}
-.navigation a:hover {
-	text-decoration: underline;
-}
-.navigation > ul > li > :first-child,
-.navigation > ul > li > :first-child * {
-	text-decoration: none !important;
-	line-height: 2rem;
-	font-family: 'Montserrat', sans-serif;
-	font-weight: 600;
-	color: var(--darkColour);
-	font-size: 1.3rem;
-}
-.navigation > ul > li > :first-child .navigation__caret {
-	visibility: hidden;
-}
-.navigation > ul > li > .navigation__item.opened {
-	color: var(--colour);
-	color: rgb(41, 117, 209);
-}
-
-.navigation__container .mobile__menu {
-	display: none;
-	height: 48px;
-	width: 48px;
-}
-@media (max-width: 500px) {
-	.navigation__container header {
-		background: transparent;
-	}
-	.navigation__container {
-		margin-right: 1rem;
-	}
-}
