--- source/sites/mon-entreprise.fr/layout/Navigation/SideBar.css
@@ -1,59 +1,0 @@
-.sidebar__container {
-	height: 100%;
-	position: absolute;
-	left: 0;
-	top: 0;
-	transform: translate(-100%);
-	/* let the browser know we plan to animate
-	   each view in and out */
-	will-change: transform;
-	transition: transform 0.3s cubic-bezier(0.465, 0.183, 0.153, 0.946);
-	z-index: 1;
-}
-.sidebar__container.opened {
-	transform: translate(0%);
-}
-
-.sidebar__container + * {
-	will-change: transform;
-	transition: transform 0.3s cubic-bezier(0.465, 0.183, 0.153, 0.946);
-}
-.sidebar__container.opened + * {
-	transform: translateX(100%);
-	overflow: hidden;
-}
-.menu__button {
-	position: absolute;
-	top: 0;
-	height: 2.5rem;
-	padding: 0.6rem;
-	width: 2.5rem;
-	right: calc(-2.5rem);
-	background: radial-gradient(
-		circle,
-		rgba(255, 255, 255, 1) 50%,
-		rgba(255, 255, 255, 0) 70%
-	);
-}
-.menu__button:hover {
-	opacity: 0.8;
-}
-.menu__button img {
-	height: 100%;
-}
-.sidebar {
-	overflow: auto;
-	height: 100%;
-}
-
-@media (min-width: 500px) {
-	.sidebar__container.opened + * {
-		transform: none;
-	}
-}
-@media (max-width: 1500px) {
-	.sidebar__container.opened + * {
-		opacity: 0.5;
-		pointer-events: none;
-	}
-}
